% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AU_engineering.R
\name{au_discrete_engineering}
\alias{au_discrete_engineering}
\title{Discrete features engineering.}
\usage{
au_discrete_engineering(df)
}
\arguments{
\item{df}{the dataframe from which extracting the facial features.}
}
\value{
a dataframe including measures of smiles, blinks, face smiles and overall count of AUCs.
}
\description{
A summary function calculating the most common discrete elements.
}
\note{
DRAFT ONLY
}
\examples{
au_discrete_engineering(john)

}
