% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_face_eyes.R
\name{draw_coordinates}
\alias{draw_coordinates}
\title{draw_coordinates
function with two options to draw a face or eyes at a sepcific frame or time}
\usage{
draw_coordinates(df, f = NULL, t = NULL, region = "face")
}
\description{
draw_coordinates
function with two options to draw a face or eyes at a sepcific frame or time
}
