% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyface.R
\name{event_summarise}
\alias{event_summarise}
\title{event summarise
summarise all the frequency variables}
\usage{
event_summarise(x, as = "count", ...)
}
\arguments{
\item{x}{the dataframe to summarise}

\item{as}{"eps", standard, is event per second. "count" is the number of events. "}
}
\description{
event summarise
summarise all the frequency variables
}
