% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{feats_in_time}
\alias{feats_in_time}
\title{features in time
visualise the level of a feature in one face, across time.}
\usage{
feats_in_time(df, vars, tmin = NULL, tmax = NULL, standardise = F,
  ...)
}
\arguments{
\item{df}{the dataframe to visualise}

\item{vars}{the variable(s) to visualise}
}
\description{
features in time
visualise the level of a feature in one face, across time.
}
