% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks finding.R
\name{findvalleys}
\alias{findvalleys}
\title{Find valleys.}
\usage{
findvalleys(x, ...)
}
\arguments{
\item{x}{a numerical array.}
}
\description{
This is the inverse of peak finding, and finds negative peaks.
}
