% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large lists.R
\name{large_lists}
\alias{large_lists}
\alias{read_and_tidy}
\alias{ll_read_of_csvs}
\alias{ll_tidyface}
\title{Dealing with large collections of csvs}
\usage{
read_and_tidy(output_dir, model)

ll_read_of_csvs(output_dir, file = "my_face_list.llo")

ll_tidyface(df, model, size = 10)
}
\arguments{
\item{output_dir}{There location where csvs files are stored.}

\item{model}{A string with the unparsed openfacer model.}

\item{file}{the name of the large list file (.llo) that will be created in your directory. Only for *ll_read_of_csv*.}

\item{df.}{The list file on which *ll_tidyface* works}

\item{c}{the size batches for *ll_tidyface*. A larger number will make the process faster but increase the
probabilities of memory failure. Default is 10.}
}
\description{
Dealing with large collections of csvs
}
\examples{
# Create a model

my.model <-  "filter_faces(success == 1) \%>\%
select_faces(AU02_r:AU45_r) \%>\%
 tidy_face(sd = FALSE)"

 # With read_and_tidy

 myface_summary <- read_and_tidy("my_file_dir/", model = my.model)

 # With large lists

 my_face_list <- ll_read_of_csv("my_file_dir/", file = "my_face_list.llo")
 myface_summary <- ll_tidyface(df = my_face_list, model = my.model, c = 10)

}
