% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEI.R
\name{MEI}
\alias{MEI}
\alias{mei_single}
\alias{mei_faces}
\alias{mei}
\title{Motion Energy Intensity}
\usage{
mei_single(df)

mei_faces(df, ...)

mei(df, ...)
}
\arguments{
\item{df:}{Either a dataframe coming from openFace output (for single faces) or an object of class "faces" (for multiple faces).}
}
\value{
A vector with one motion energy per face.
}
\description{
Motion Energy Intensity
}
\details{
The function calculate the average energy of motion of 67 the facial points.The frame to frame difference in movement of each facial point is calculated after correcting for the average position of the most rigid points (nose bridge and face sides). The change of each point is summed for each frame and the resulting vector is averaged.
}
\note{
This is still a draft and the final version will use a different algorithm
}
\examples{

# Calculate Lara frame to frame mei
mei(lara)

# Calculate average mei for the four example videos
test_videos \%>\%
  transform_videos("mei", mei) \%>\%
  select_videos(mei) \%>\%
  tidy_face()

}
