% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks finding.R
\name{moving_av}
\alias{moving_av}
\title{Moving average.}
\usage{
moving_av(x, smooth)
}
\arguments{
\item{x}{a numerical array.}

\item{smooth}{the smooth parameter is used to smooth the descriptive curve and avoid the inclusion of very small differences in values. The default (advised) is 10.}
}
\description{
The function calculate the moving average of an array of numbers and add NAs in the end to have the same number of arguments.
}
