% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{moving_difference}
\alias{moving_difference}
\title{Calculate velocity.
calculate the difference of values in each ordered value of an array}
\usage{
moving_difference(x)
}
\arguments{
\item{x}{The array to transform}
}
\value{
an array of same length (first value is NA)
}
\description{
Calculate velocity.
calculate the difference of values in each ordered value of an array
}
\examples{
moving_difference(c(1,4,-2, 2))
}
