% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks finding.R
\name{peaks}
\alias{peaks}
\alias{peaks_stats}
\alias{nvb_peaks}
\title{Peaks.}
\usage{
peaks_stats(df, var = NULL, smooth = 10, thresh = 5)

nvb_peaks(df, var, smooth = 10, thresh = 5)
}
\arguments{
\item{df}{the dataframe in which the variables are stored or a dataframe of peaks values.}

\item{var}{the target variable in the dataframe.}

\item{smooth}{the smooth parameter is used to smooth the descriptive curve and avoid the inclusion of very small differences in values. The default (advised) is 10.}

\item{tresh}{it indicates the minimum threshold to use to determine peaks used by Pracma "findPeaks" function. The  default (5) is advised.}
}
\description{
Set of functions to record statistics about peaks of movements. Especially used to identify distinct movements of the head.
}
\examples{
# Describes the frequency of Teddy moving the head horizontaly
peaks_stats(teddy, var = pose_Tx, smooth = 10, thresh = 5)

# Find the peaks of horizontal head motion for Lara
nvb_peaks(lara, var = pose_Tx, smooth = 10, thresh = 5)

# Find Lara peaks and sumarise them (with conservative parameters)
lara_peaks <- nvb_peaks(lara, var = pose_Tx, smooth = 20, thresh = 10)
peaks_stats(lara_peaks)
}
