% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyface.R
\name{tidy_face}
\alias{tidy_face}
\title{tidy_face
A core function of the OpenFaceR routine. It summarise all the columns of a dataframe into a tidy dataset}
\usage{
tidy_face(x, events = TRUE, continuous = TRUE, events_sum = "count",
  ...)
}
\arguments{
\item{...}{more options to control the summary of continuous variables (see continuous_summarise)}

\item{x:}{a faces object}

\item{events:}{include events variables summary}

\item{continuous:}{include countinous variables summary}

\item{events_sum:}{the method for summarising events. This can be "eps" (events per second), "count" or "rate", which is the proportion of time in which events are happening}
}
\description{
tidy_face
A core function of the OpenFaceR routine. It summarise all the columns of a dataframe into a tidy dataset
}
