% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyface.R
\name{time_summarise}
\alias{time_summarise}
\title{time summarise
The other information we want to keep is the total time of the video}
\usage{
time_summarise(x)
}
\arguments{
\item{x:}{a faces dataframe}
}
\value{
a dataframe with two variables: the video ID and the length in seconds of each video
}
\description{
time summarise
The other information we want to keep is the total time of the video
}
