% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mutate_select_tidy.R
\name{transform_faces}
\alias{transform_faces}
\title{Transform
unlike mutate it takes function taking as argument the full dataset rather than dataset variables}
\usage{
transform_faces(faces, var, fun)
}
\description{
Transform
unlike mutate it takes function taking as argument the full dataset rather than dataset variables
}
