% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head moves.R
\name{velocity_3d}
\alias{velocity_3d}
\title{Head Movements.}
\usage{
velocity_3d(df)
}
\arguments{
\item{df}{A dataframe with facial information from OpenFace.}
}
\value{
A frame by frame array with the total difference of spatial position.
}
\description{
Calculates the overall amount of head movements in the three axis.
}
\examples{

# Calculate Lara frame to frame change in head position
velocity_3d(lara)

# Calculate average velocity of head for the four example videos
test_videos \%>\%
  transform_videos("velocity", velocity_3d) \%>\%
  select_videos(velocity_3d) \%>\%
  tidy_face()

}
