% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{count_events}
\alias{count_events}
\title{count events
Function to count howm many differnet time a facial event occurs}
\usage{
count_events(x, sensitivity = 5)
}
\arguments{
\item{x}{the variable for which to count events (must have two levels: 0 and 1)}

\item{sensitivity}{how many frames might have the sodtware missed and still considered the event as one}
}
\description{
count events
Function to count howm many differnet time a facial event occurs
}
