% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical.R
\name{feats_in_time}
\alias{feats_in_time}
\title{features in time
visualise the level of a feature in one face, across time.}
\usage{
feats_in_time(df, vars, tmin = NULL, tmax = NULL, standardise = F,
  ...)
}
\arguments{
\item{df}{The dataframe to visualise.}

\item{vars}{The variable(s) to visualise. There is a list of preset functions:
"head" or "h" allows to visualise the movements of the head in the three spatial axis.}

\item{tmin}{Starting time}

\item{tmax}{Ending time}

\item{standardise}{If true, standardises all the variables (to use when comparing variables with different levels)}
}
\value{
A plot with the level of one or more features in time
}
\description{
features in time
visualise the level of a feature in one face, across time.
}
\examples{

# see the two AU relative to smiles for John, from second 10 to second 30
feats_in_time(john, c("AU_06r", "AU_12r"), tmin = 10, tmax = 30, standardise = F)

# Paula standardised head movements
feats_in_time(paula, "h", standardise = TRUE)
}
