% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string for extraction creation.R
\name{get_commands}
\alias{get_commands}
\title{get command
provides two strings to consecutively use n the control pane to get videos features extracted into csv format}
\usage{
get_commands(of_dir, input_dir, output_dir, filename = NULL)
}
\arguments{
\item{:}{of_dir: the folder where openface executable is stored}

\item{:}{input_dir: the folder where to save csv files}

\item{filename:}{file name to provide if the goal is to only analyse one video}
}
\description{
get command
provides two strings to consecutively use n the control pane to get videos features extracted into csv format
}
