% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_of_csv.R
\name{read_face_csvs}
\alias{read_face_csvs}
\title{read face csvs
Function to read all the csvs from one analysis into one face object}
\usage{
read_face_csvs(output_dir)
}
\arguments{
\item{:}{output_dir the folder where csvs outputs are stored}
}
\description{
read face csvs
Function to read all the csvs from one analysis into one face object
}
