% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head moves.R
\name{velocity_3d}
\alias{velocity_3d}
\title{head movements
three dimensional velocity (measure of overall head movement)}
\usage{
velocity_3d(df)
}
\description{
head movements
three dimensional velocity (measure of overall head movement)
}
